refactor(network-api): GatewayList 구현체 'ip route show'로 변경
