+++ modules/network-api/src/main/java/com/whoz_in/network_api/common/network_interface/IwconfigNetworkInterfaces.java
@@ -1,4 +1,4 @@
-package com.whoz_in.network_api.common;
+package com.whoz_in.network_api.common.network_interface;
 
 import com.whoz_in.network_api.common.process.TransientProcess;
 import java.util.ArrayList;
