+++ modules/network-api/src/main/java/com/whoz_in/network_api/config/NetworkConfig.java
@@ -1,7 +1,6 @@
 package com.whoz_in.network_api.config;
 
-import com.whoz_in.network_api.common.NetworkInterface;
-import com.whoz_in.network_api.config.NetworkInterfaceProperties.Mdns;
+import com.whoz_in.network_api.common.network_interface.NetworkInterface;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
