+++ modules/network-api/src/main/java/com/whoz_in/network_api/managed/mdns/MdnsLogWriter.java
@@ -2,8 +2,8 @@ package com.whoz_in.network_api.managed.mdns;
 
 import com.whoz_in.domain.network_log.ManagedLog;
 import com.whoz_in.domain.network_log.ManagedLogRepository;
-import com.whoz_in.network_api.common.NetworkInterface;
-import com.whoz_in.network_api.common.SystemNetworkInterfaces;
+import com.whoz_in.network_api.common.network_interface.NetworkInterface;
+import com.whoz_in.network_api.common.network_interface.SystemNetworkInterfaces;
 import com.whoz_in.network_api.common.process.ContinuousProcess;
 import com.whoz_in.network_api.config.NetworkConfig;
 import com.whoz_in.network_api.managed.ParsedLog;
