+++ modules/network-api/src/main/java/com/whoz_in/network_api/private_ip/PrivateIpUpdater.java
@@ -1,6 +1,6 @@
 package com.whoz_in.network_api.private_ip;
 
-import com.whoz_in.network_api.common.NetworkInterface;
+import com.whoz_in.network_api.common.network_interface.NetworkInterface;
 import com.whoz_in.network_api.common.util.SystemPrivateIpResolver;
 import com.whoz_in.network_api.config.NetworkConfig;
 import java.util.HashMap;
