+++ modules/network-api/src/main/java/com/whoz_in/network_api/system_validator/MonitorModeSwitcher.java
@@ -1,6 +1,6 @@
 package com.whoz_in.network_api.system_validator;
 
-import com.whoz_in.network_api.common.NetworkInterface;
+import com.whoz_in.network_api.common.network_interface.NetworkInterface;
 import com.whoz_in.network_api.common.process.TransientProcess;
 import com.whoz_in.network_api.config.NetworkConfig;
 import lombok.extern.slf4j.Slf4j;
